+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * Implement SDL grab functions
 * Add ImPlot and Unity job count debug window
 * Prevent games from knowing that game window is hidden or minimized
+* More checks and display errors on Ram Search
 
 ### Changed